data-source/app/requirements.txt
kafka-python
requests
shortuuid